figured out how to dynamically display the title of the current page
I used php because it is more reliable (e.g. I don't have to rely on the client having JS enabled) than JS, and it comes with the added benefit of generating HTML before the browser ever has a chance of displaying it (no chance for a brief flash of placeholder text while the JS executes).
I still need to clean up this file and figure out how to actually use the text I found.